Initial work to fix the big bang issue (#8)
* Initial work to fix the big bang issue

* Fixes for AWS API

* Fixing API calls (not working as documented)

* Fixing typing import

+++ joule/providers/aws.py
@@ -4,9 +4,10 @@ import logging
 
 from datetime import datetime
 from ec2_metadata import ec2_metadata
-from typing import Iterator, Optional
+from typing import Iterator, Optional, List, Dict
 
 from mypy_boto3_autoscaling.client import AutoScalingClient
+from mypy_boto3_ec2 import EC2Client
 from mypy_boto3_sqs.service_resource import Message, Queue
 
 from joule.providers import BaseProvider, Event, Events, logging
@@ -27,7 +28,7 @@ class AwsProvider(BaseProvider):
 
         super().__init__(*applications)
 
-        self.instance_id: str = ec2_metadata.instance_id
+        self._instance_id: str = ec2_metadata.instance_id
 
         self._region: str = ec2_metadata.region
 
@@ -36,11 +37,12 @@ class AwsProvider(BaseProvider):
         ][
             0
         ]  # Get first queue found.
-        self.queue: Queue = boto3.resource("sqs", region_name=self._region).Queue(
+        self._queue: Queue = boto3.resource("sqs", region_name=self._region).Queue(
             queue_url
         )
 
-        self.asg: AutoScalingClient = boto3.client(
+        self._ec2: EC2Client = boto3.client("ec2", region_name=self._region)
+        self._asg: AutoScalingClient = boto3.client(
             "autoscaling", region_name=self._region
         )
 
@@ -54,18 +56,52 @@ class AwsProvider(BaseProvider):
         if datetime.now().second != 0:
             return  # Run every minute.
 
-        asg_name: str = self.asg.describe_auto_scaling_instances(
+        asg_name: str = self._asg.describe_auto_scaling_instances(
             InstanceIds=[ec2_metadata.instance_id], MaxRecords=1
         )["AutoScalingInstances"][0]["AutoScalingGroupName"]
 
         for app in self.applications:
             is_essential: bool = app.is_essential()
-            self.asg.set_instance_protection(
+            self._asg.set_instance_protection(
                 InstanceIds=[ec2_metadata.instance_id],
                 AutoScalingGroupName=asg_name,
                 ProtectedFromScaleIn=is_essential,
             )
             logging.debug("essential={}".format(is_essential))
+
+    def mark_enrolled(self) -> None:
+        """
+        Mark this instance as being enrolled in the application's cluster.
+
+        :return: None
+        """
+        self._ec2.create_tags(
+            Resources=[self._instance_id],
+            Tags=[
+                {"Key": self._tag_enrolled["Key"], "Value": self._tag_enrolled["Value"]}
+            ],
+        )
+
+    def is_enrolled(self) -> bool:
+        """
+        Return whether or not this instance has been enrolled into the
+        application's cluster.
+
+        :return: Boolean
+        """
+        result: dict = self._ec2.describe_tags(
+            Filters=[
+                {"Name": "resource-id", "Values": [self._instance_id]},
+                {
+                    "Name": "key",
+                    "Values": [self._tag_enrolled["Key"]],
+                },
+            ],
+        )
+
+        if result.get("Tags", None):
+            return True
+        return False  # When none or empty list (no matches for tag key).
 
     def get_events_from_message_queue(self) -> Iterator[Optional[Event]]:
         """
@@ -73,25 +109,30 @@ class AwsProvider(BaseProvider):
 
         :return: Event
         """
-        rx: list[Message] = self.queue.receive_messages()
+        rx: List[Message] = self._queue.receive_messages()
 
         for msg in rx:
             try:
-                loaded: dict[str, str] = json.loads(json.loads(msg.body).get("Message"))
+                loaded: Dict[str, str] = json.loads(
+                    json.loads(msg.body).get("Message")
+                )
             except TypeError:
-                loaded: dict[str, str] = json.loads(msg.body)
+                loaded: Dict[str, str] = json.loads(msg.body)
 
-            if loaded.get("Event") == "autoscaling:EC2_INSTANCE_LAUNCH":
-                msg.delete()
-                yield Event(event=Events.LAUNCH, instance=loaded["EC2InstanceId"])
+            if self.is_enrolled():
+                if loaded.get("Event") == "autoscaling:EC2_INSTANCE_LAUNCH":
+                    msg.delete()
+                    yield Event(event=Events.LAUNCH, instance=loaded["EC2InstanceId"])
 
-            if loaded.get("Event") == "autoscaling:EC2_INSTANCE_TERMINATE":
-                msg.delete()
-                yield Event(event=Events.TERMINATE, instance=loaded["EC2InstanceId"])
+                if loaded.get("Event") == "autoscaling:EC2_INSTANCE_TERMINATE":
+                    msg.delete()
+                    yield Event(
+                        event=Events.TERMINATE, instance=loaded["EC2InstanceId"]
+                    )
 
             for app in self.applications:
                 if loaded.get("Event") == "{}:join".format(app.name):
-                    if loaded.get("EC2InstanceId") == self.instance_id:
+                    if loaded.get("EC2InstanceId") == self._instance_id:
                         msg.delete()
                         yield Event(
                             event=Events.JOIN,
@@ -113,7 +154,7 @@ class AwsProvider(BaseProvider):
         :param payload: Dictionary
         :return: None
         """
-        self.queue.send_message(
+        self._queue.send_message(
             MessageBody=json.dumps(
                 {
                     "Event": "{}:join".format(application.name),
